[REFACTOR] Moved paramaters
The group is now constructed with a setup class. This moves the
construction of the parameters to the setup class.
