+++ core/src/main/java/org/mammon/math/Gq.java
@@ -1,34 +1,19 @@
 package org.mammon.math;
 
 import java.math.BigInteger;
-import java.util.List;
 
 import org.mammon.math.util.Gcd;
-import org.mammon.math.util.PrimeFactors;
 
 public class Gq implements Group<Gq> {
 
 	private final BigInteger q;
 	private final BigInteger p;
-	private Group.Element<Gq> primitiveRoot;
-	private List<BigInteger> factorsOftotientP;
-	private Group.Element<Gq> generator;
+	private final Group.Element<Gq> generator;
 
-	public Gq(int q, int p) {
-		this(BigInteger.valueOf(q), BigInteger.valueOf(p));
-	}
-
-	public Gq(BigInteger q, BigInteger p) {
-		qShouldDivideTotientP(q, p);
-		this.q = q;
-		this.p = p;
-	}
-
-	private void qShouldDivideTotientP(BigInteger q, BigInteger p) {
-		BigInteger[] divmod = p.subtract(BigInteger.ONE).divideAndRemainder(q);
-		if (!divmod[1].equals(BigInteger.ZERO)) {
-			throw new IllegalArgumentException("q should divide p-1");
-		}
+	public Gq(GqSetup setup) {
+		this.q = setup.q;
+		this.p = setup.p;
+		this.generator = element(setup.generator);
 	}
 
 	@Override
@@ -42,46 +27,7 @@ public class Gq implements Group<Gq> {
 
 	@Override
 	public Group.Element<Gq> getGenerator() {
-		if (generator == null) {
-			Z field = new Z(p);
-			Group.Element<Gq> primitiveRoot = getPrimitiveRoot();
-			generator = primitiveRoot.exponentiate(field.element(p.subtract(BigInteger.ONE)).multiply(
-				field.element(q).getInverse()));
-		}
 		return generator;
-	}
-
-	private Group.Element<Gq> getPrimitiveRoot() {
-		if (primitiveRoot == null) {
-			primitiveRoot = generatePrimitiveRoot();
-		}
-		return primitiveRoot;
-	}
-
-	private Group.Element<Gq> generatePrimitiveRoot() {
-		BigInteger candidate = BigInteger.valueOf(2);
-		while (!isPrimitiveRoot(candidate)) {
-			candidate = candidate.add(BigInteger.ONE);
-		}
-		return element(candidate);
-	}
-
-	private boolean isPrimitiveRoot(BigInteger candidate) {
-		List<BigInteger> factors = factorsOfTotientP();
-		for (BigInteger factor : factors) {
-			if (p.mod(factor).equals(BigInteger.ZERO)) {
-				return false;
-			}
-		}
-		return true;
-	}
-
-	private List<BigInteger> factorsOfTotientP() {
-		if (factorsOftotientP == null) {
-			factorsOftotientP = PrimeFactors.of(p.subtract(BigInteger.ONE));
-
-		}
-		return factorsOftotientP;
 	}
 
 	@Override
